Refactor Snapshot Strip in SpatialDisplayComponent.vue
- Replaced the existing snapshot strip implementation with a new SnapshotStrip component for improved modularity and maintainability.
- Removed redundant snapshot-related state management and calculations from the SpatialDisplayComponent.
- Updated event handling for snapshot selection and closing to utilize the new component's emitted events.
- Cleaned up unused imports and styles related to the old snapshot strip implementation.
